refactor(): a lot of changes tweaks & improvements everywhere !
