+++ src/api/ops-measurements/bounds.js
@@ -1,10 +1,10 @@
-const cagMeasurements = require('../../core/utils/cagMeasurements')
-const csgMeasurements = require('../../core/utils/csgMeasurements')
+const measureBounds2d = require('../../core/geometry/shape2/measureBounds')
+const measureBounds3d = require('../../core/geometry/shape3/measureBounds')
 
-const {isCAG} = require('../../core/utils')
+const {isShape2} = require('../../core/utils/typeChecks')
 
 const bounds = input => {
-  return isCAG(input) ? cagMeasurements.getBounds(input) : csgMeasurements.bounds(input)
+  return isShape2(input) ? measureBounds2d(input) : measureBounds3d(input)
 }
 
 module.exports = bounds
