+++ src/core/geometry/side/direction.js
@@ -1,0 +1,3 @@
+direction: function () {
+  return this.vertex1.pos.minus(this.vertex0.pos)
+},
