+++ src/core/geometry/side/flip.js
@@ -1,0 +1,3 @@
+flipped: function () {
+  return new Side(this.vertex1, this.vertex0)
+},
