+++ src/core/geometry/side/fromFakePolygon.js
@@ -1,0 +1,31 @@
+Side._fromFakePolygon = function (polygon) {
+  // this can happen based on union, seems to be residuals -
+  // return null and handle in caller
+if (polygon.vertices.length < 4) {
+  return null
+}
+var vert1Indices = []
+var pts2d = polygon.vertices.filter(function (v, i) {
+  if (v.pos.z > 0) {
+    vert1Indices.push(i)
+    return true
+  }
+  return false
+})
+  .map(function (v) {
+    return new Vector2D(v.pos.x, v.pos.y)
+  })
+if (pts2d.length !== 2) {
+  throw new Error('Assertion failed: _fromFakePolygon: not enough points found')
+}
+var d = vert1Indices[1] - vert1Indices[0]
+if (d === 1 || d === 3) {
+  if (d === 1) {
+    pts2d.reverse()
+  }
+} else {
+  throw new Error('Assertion failed: _fromFakePolygon: unknown index ordering')
+}
+var result = new Side(new Vertex(pts2d[0]), new Vertex(pts2d[1]))
+return result
+}
