+++ src/core/geometry/side/fromObject.js
@@ -1,0 +1,5 @@
+Side.fromObject = function (obj) {
+  var vertex0 = Vertex.fromObject(obj.vertex0)
+  var vertex1 = Vertex.fromObject(obj.vertex1)
+  return new Side(vertex0, vertex1)
+}
