+++ src/core/geometry/side/length.js
@@ -1,0 +1,7 @@
+const lengthSquared = require('./lengthSquared')
+
+const length = side => {
+  return Math.sqrt(lengthSquared(side))
+}
+
+module.exports = length
