+++ src/core/geometry/side/toString.js
@@ -1,0 +1,3 @@
+toString: function () {
+  return this.vertex0 + ' -> ' + this.vertex1
+},
